chore: remove expo-router dependencies from non-expo test templates
