+++ integration-tests/templates/expo-rn74-prebuild-template/App.tsx
@@ -1,11 +1,10 @@
 import React from 'react';
-import {EmbraceTestHarness} from '@embrace-io/react-native-test-harness';
+import {EmbraceExpoTestHarness} from '@embrace-io/react-native-test-harness/lib/EmbraceExpoTestHarness';
 import sdkConfig from './embrace-sdk-config.json';
 
 function App(): React.JSX.Element {
   return (
-    <EmbraceTestHarness
-      navigationStyle="react-native"
+    <EmbraceExpoTestHarness
       sdkConfig={sdkConfig}
       allowCustomExport={false}
     />
