+++ integration-tests/templates/expo-rn74-template/app/_layout.tsx
@@ -1,14 +1,10 @@
 import React from "react";
 import "react-native-reanimated";
-import {EmbraceTestHarness} from "@embrace-io/react-native-test-harness";
+import {EmbraceExpoTestHarness} from "@embrace-io/react-native-test-harness/lib/EmbraceExpoTestHarness";
 import sdkConfig from "./embrace-sdk-config.json";
 
 export default function RootLayout() {
   return (
-    <EmbraceTestHarness
-      navigationStyle="expo"
-      sdkConfig={sdkConfig}
-      allowCustomExport={false}
-    />
+    <EmbraceExpoTestHarness sdkConfig={sdkConfig} allowCustomExport={false} />
   );
 }
