+++ integration-tests/templates/rn72-template/package.json
@@ -15,8 +15,7 @@
     "react": "18.2.0",
     "react-native": "0.72.17",
     "react-native-safe-area-context": "4.11.1",
-    "react-native-screens": "3.34.0",
-    "expo-router": "3.5.14"
+    "react-native-screens": "3.34.0"
   },
   "devDependencies": {
     "@babel/core": "7.20.2",
