+++ integration-tests/templates/rn74-template/App.tsx
@@ -1,11 +1,10 @@
 import React from 'react';
-import {EmbraceTestHarness} from '@embrace-io/react-native-test-harness';
+import {EmbraceReactNativeTestHarness} from '@embrace-io/react-native-test-harness';
 import sdkConfig from './embrace-sdk-config.json';
 
 function App(): React.JSX.Element {
   return (
-    <EmbraceTestHarness
-      navigationStyle="react-native"
+    <EmbraceReactNativeTestHarness
       sdkConfig={sdkConfig}
       allowCustomExport={false}
     />
