+++ integration-tests/templates/rn76-template/package.json
@@ -15,8 +15,7 @@
     "react": "18.3.1",
     "react-native": "0.76.5",
     "react-native-safe-area-context": "5.0.0",
-    "react-native-screens": "4.4.0",
-    "expo-router": "3.5.14"
+    "react-native-screens": "4.4.0"
   },
   "devDependencies": {
     "@babel/core": "7.25.2",
