+++ integration-tests/templates/rn79-template/package.json
@@ -15,8 +15,7 @@
     "react": "19.0.0",
     "react-native": "0.79.1",
     "react-native-safe-area-context": "5.5.2",
-    "react-native-screens": "4.9.2",
-    "expo-router": "3.5.14"
+    "react-native-screens": "4.9.2"
   },
   "devDependencies": {
     "@babel/core": "7.25.2",
